bull_project/requirements.txt
aiofiles==24.1.0
aiogram==3.22.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
aiosqlite==0.21.0
asyncpg==0.30.0
annotated-types==0.7.0
attrs==25.4.0
cachetools==6.2.2
certifi==2025.11.12
charset-normalizer==3.4.4
decorator==5.2.1
easyocr==1.7.2
filelock==3.20.0
frozenlist==1.8.0
fsspec==2025.12.0
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.41.1
google-auth-httplib2==0.2.1
google-auth-oauthlib==1.2.3
googleapis-common-protos==1.72.0
greenlet==3.2.4
gspread==6.2.1
httplib2==0.31.0
idna==3.11
ImageIO==2.37.2
Jinja2==3.1.6
lazy_loader==0.4
magic-filter==1.0.12
MarkupSafe==3.0.3
mpmath==1.3.0
mrz==0.6.2
multidict==6.7.0
networkx==3.6
ninja==1.13.0
numpy==2.2.6
oauthlib==3.3.1
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
packaging==25.0
pdf2image==1.17.0
pillow==12.0.0
propcache==0.4.1
proto-plus==1.26.1
protobuf==6.33.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyclipper==1.4.0
pydantic==2.11.10
pydantic_core==2.33.2
pyparsing==3.2.5
pytesseract==0.3.13
python-bidi==0.6.7
PyYAML==6.0.3
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
scikit-image==0.25.2
scipy==1.16.3
self==2020.12.3
shapely==2.1.2
SQLAlchemy==2.0.44
sympy==1.14.0
tifffile==2025.10.16
torch==2.9.1
torchvision==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.5.0
yarl==1.22.0
python-multipart
fastapi
uvicorn[standard]
